Refactored damagecalc.js code into _damagecalc.js. Added to-do-list.
I've recreated the program's architecture, using a more object oriented approach to the problem in hand. I separated it into three main objects - pokemonBattle, battleModifiers and validator - that encapsulates all of the necessary functions for a damage calculator. Also, I realized today that pokemonBattle can be improved and I'll do it after the beta release of the calc, i.e. when I've implemented all of the variables involved.