Rearchitectured calculatorModel.calcDamage.
Added a string input so you can say if you want a minimal or maximal damage, instead of creating two methods. Also, I want to implement some way to pass a value parameter instead of (only) a string, so you can select the desired multiplier. This way, it'll be possible to know the probability that a move will faint a pokemon, not just the min and max damage.